Add action tab, a command engine, and virtual sink creation.
This commit adds another tab to the main window that lets the user
create a new virtual sink by supplying a name. It's rough around the
edges (no feedback on completion, no name collision testing), but a step
in the right direction.

More importantly, this commit establishes a framework for adding new
commands.