Split off command tab from action tab, and redo dialogs.
When a tree representing an action is selected, the tab now shows an
editor window containing the action's information. This is, I think, far
more useful than trying to representing all the data in a command as a
series of tree items.

The action tab now focuses on immediate execution, whereas the new
command tab is for user-defined commands.