Fix commands that create a sink and connect it failing.
Prior to this commit, commands were executed by sending them to the
execution engine and immediately running them. That strategy worked most
of the time. However, commands that wanted to create sinks and then
route to them would fail since the newly-made sinks wouldn't be visible
in `ZeusPulseData`.

With this commit, the command tab gains a `ZeusPulseContext`. The
context will execute a command until an action specifies a sink to await
for. When the context gets the signal that the sink is connected,
command execution continues. This overall makes commands much more
robust than they were before.