Playback and record tabs now report move failures.
In most cases, moving a stream to a different device is fine. However,
some streams are created with `PA_STREAM_DONT_MOVE`. For those streams,
PulseAudio will refuse to move the stream.

It appears to be possible to detect if a stream is set to DONT_MOVE by
attempting to move it to the device that it's currently on. However, I'd
rather not have Zeus create unnecessary traffic to detect a situation
that isn't common.

Move failures from actions are still not reported, because I'm not yet
sure if/how I want to show those errors.