`ZeusDeviceComboBox` now supports placeholders, fixing a couple issues.
The first issue is that some active PulseAudio streams will have an
invalid device (`-1`). Zeus wasn't accounting for that, and so those
streams in the Playback / Record tab would show the wrong device. They
now show a placeholder item that has `Unknown Device`.

The second issue was with Zeus showing saved actions for commands. If an
action used a device that wasn't available when being previewed, Zeus
would show a different device. The issue was purely with representation,
because once the devices became available, the preview was showing the
correct values.

With this commit, those saved actions now show the correct name in the
box, resulting in less user confusion. :)