Adjust json config format for auto sorting.
Before:

```
    "commands": [
        {
            "name": ...
            "actions": [ ... ]
        }
    ]
```

After:

```
    "commands": [
        <command name>: [ ... ]
    ]
```

This isn't compatible with prior configs, but that's fine since I'm the
only user.

I'm doing this because I keep my Zeus config in source control. Since Qt
sorts object keys when producing json output, this should result in
fewer "changes" where the data is the same but the (command) keys are
in a different order.